Revamp of the binding (and the reporting) layer.
Follow the process binding provided by the batch scheduler or process manager.
If the application is started with mpirun, follow the bindings provided by the
mpirun command.

Add an MCA parameter (runtime_report_bindings) to report the final bindings for
each virtual process and thread. Report the binding in both logical and
physical notation.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>