Increment the data epoch everytime we call zone_free
We now call `zone_free` in places that don't update the data epoch.
It's important to signal that memory is available again because otherwise
we get stuck. There is a potential race condition when the update happen
in `parsec_data_discard` but the exact value of the epoch is irrelevant,
only that the epoch has changed.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>